+++ src/Core/LibraryApp.Application/Feauters/Reviews/Commands/Update/UpdateReviewCommand.cs
@@ -4,6 +4,6 @@ namespace LibraryApp.Application.Feauters.Reviews.Commands.Update;
 
 public record UpdateReviewCommand(
 	Guid ReviewId,
-	int Rating,
+	byte Rating,
 	string? Title,
 	string? Text) : IRequest<Unit>;
